Adds ability to specify and use a digest for base image ref
The triggering issue for this change is that apparently Docker Stacks
upgraded Python from 3.11 to 3.12 for their current series of image tags
(all the way up from their base image). This in turn prevents
tensorflow-text from installing. (I suppose it requires Python <= 3.11)

So for the time being we resort to using the SHA digests that were most
recently working. (These may not be the most recent with Python 3.11 inside)